clean: Move TotalComputationResources() related code to v7 pkg only
