+++ vm/trace_test.go
@@ -138,96 +138,6 @@ func TestAllMessages(t *testing.T) {
 	}
 }
 
-func TestTotalExecutionResources(t *testing.T) {
-	resources := &vm.ExecutionResources{
-		ComputationResources: vm.ComputationResources{
-			Steps:        1,
-			MemoryHoles:  2,
-			Pedersen:     3,
-			RangeCheck:   4,
-			Bitwise:      5,
-			Ecdsa:        6,
-			EcOp:         7,
-			Keccak:       8,
-			Poseidon:     9,
-			SegmentArena: 10,
-		},
-	}
-	tests := map[string]struct {
-		multiplier uint64
-		trace      *vm.TransactionTrace
-	}{
-		"many top-level invocations": {
-			multiplier: 5,
-			trace: &vm.TransactionTrace{
-				ValidateInvocation: &vm.FunctionInvocation{
-					ExecutionResources: resources,
-				},
-				FunctionInvocation: &vm.FunctionInvocation{
-					ExecutionResources: resources,
-				},
-				ConstructorInvocation: &vm.FunctionInvocation{
-					ExecutionResources: resources,
-				},
-				ExecuteInvocation: &vm.ExecuteInvocation{
-					FunctionInvocation: &vm.FunctionInvocation{
-						ExecutionResources: resources,
-					},
-				},
-				FeeTransferInvocation: &vm.FunctionInvocation{
-					ExecutionResources: resources,
-				},
-			},
-		},
-		"only validate invocation": {
-			multiplier: 1,
-			trace: &vm.TransactionTrace{
-				ValidateInvocation: &vm.FunctionInvocation{
-					ExecutionResources: resources,
-				},
-			},
-		},
-		"present in some sub-calls": {
-			multiplier: 2,
-			trace: &vm.TransactionTrace{
-				ValidateInvocation: &vm.FunctionInvocation{
-					ExecutionResources: resources,
-					Calls: []vm.FunctionInvocation{
-						{
-							ExecutionResources: resources,
-						},
-					},
-				},
-				FunctionInvocation: &vm.FunctionInvocation{
-					ExecutionResources: resources,
-					Calls: []vm.FunctionInvocation{
-						{
-							ExecutionResources: resources,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for description, test := range tests {
-		t.Run(description, func(t *testing.T) {
-			require.Equal(t, vm.ComputationResources{
-				Steps:        resources.Steps * test.multiplier,
-				MemoryHoles:  resources.MemoryHoles * test.multiplier,
-				Pedersen:     resources.Pedersen * test.multiplier,
-				RangeCheck:   resources.RangeCheck * test.multiplier,
-				Bitwise:      resources.Bitwise * test.multiplier,
-				Ecdsa:        resources.Ecdsa * test.multiplier,
-				EcOp:         resources.EcOp * test.multiplier,
-				Keccak:       resources.Keccak * test.multiplier,
-				Poseidon:     resources.Poseidon * test.multiplier,
-				SegmentArena: resources.SegmentArena * test.multiplier,
-			}, test.trace.TotalComputationResources())
-		})
-	}
-}
-
 func TestExecuteInvocation(t *testing.T) {
 	tests := map[string]struct {
 		inv      vm.ExecuteInvocation
